pyproject.toml
[project]
name = "agentkit-sdk-python"
version = "0.3.0"
description = "Python SDK for transforming any AI agent into a production-ready application. Framework-agnostic primitives for runtime, memory, authentication, and tools with volcengine-managed infrastructure."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
license = { file = "LICENSE" }
authors = [
    {name = "Xiangrui Cheng", email = "innsdcc@gmail.com"},
    {name = "Yumeng Bao", email = "baoyumeng.123@gmail.com"},
    {name = "Yaozheng Fang", email = "fangyozheng@gmail.com"},
    {name = "Guodong Li", email = "cu.eric.lee@gmail.com"},
]
dependencies = [
    "a2a-sdk>=0.3.7",
    "fastapi>=0.117.1",
    "fastmcp>=2.12.3",
    "opentelemetry-api>=1.32.1, <=1.37.0",
    "opentelemetry-exporter-otlp-proto-common>=1.32.1, <=1.37.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.32.1, <=1.37.0",
    "opentelemetry-sdk>=1.32.1, <=1.37.0",
    "pydantic>=2.11.9",
    "requests>=2.32.5",
    "uvicorn>=0.37.0",
    "pyyaml",
    "python-dotenv>=1.1.0",
    "prompt_toolkit",
    "typer",
    "rich",
    "docker",
    "jinja2",
    "tos>=2.8.7",
    "volcengine",
    "pyfiglet>=1.0.2",
    "pyreadline3; sys_platform == 'win32'",
]

[project.scripts]
agentkit = "agentkit.toolkit.cli.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentkit*"]

[tool.setuptools]
include-package-data = true

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
]
toolkit = [
    "pyyaml>=6.0.2",
    "typer>=0.19.2",
    "inquirerpy>=0.3.4",
]

[project.optional-dependencies]
extensions = []
toolkit = []
dev = []

[tool.ruff.lint]

[tool.ruff]
line-length = 88

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "agentkit/toolkit/resources/samples/",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["F401", "F811"] 
